// src/types/payer.ts
export interface IPayerProps {
  payer_id: string;
  payer_name: string;
  payer_address: string;
  payer_address_state: string;
  payer_address_zipcode: string;
  payer_address_neighbordhood: string;
  payer_address_addressName: string;
  payer_address_number: string;
  payer_address_complement?: string;
}